# docs/contracts/v1/openapi.yaml
openapi: '3.1.1'
info:
  title: Gestão de Eventos — API
  description: "API do **monolito modular** de gestão de eventos.\n\n## Convenções\n- Prefixo `api/v1/{modulo}`; recursos no plural, sem verbos na URL.\n- Erros seguem **RFC 9457 (Problem Details)** com a extensão `codigo` (estável, ex.: `Eventos.EventoNaoEncontrado`).\n- Listagens são paginadas: `pagina` e `tamanhoPagina` (máx. 100) e retornam `itens`, `total`, `totalPaginas`.\n- Todo response carrega `X-Correlation-Id` (aceita o enviado pelo cliente) para correlação com logs e traces.\n- Autenticação **Bearer JWT**: obtenha o token em `POST api/v1/identidade/sessoes`.\n\n## Módulos\nCada módulo possui seu schema no banco, seus casos de uso e seu grupo de endpoints (uma tag por módulo)."
  contact:
    name: Globalsys
    url: https://globalsys.com.br
  version: v1
servers:
  - url: http://127.0.0.1:5762/
paths:
...
  '/api/v1/eventos/{id}/trilhas':
    get:
      tags:
        - Eventos
      summary: Lista as trilhas de um evento
      description: Retorna as trilhas temáticas ordenadas por nome; use `estaAtivo=true` para seleção em palestras.
      operationId: ListarTrilhas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: EstaAtivo
          in: query
          schema:
            type: boolean
      responses:
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListarTrilhasItemResponse'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    post:
      tags:
        - Eventos
      summary: Adiciona uma trilha temática ao evento
      description: O nome da trilha é único no evento. **Perfil exigido:** Administrador ou Organizador.
      operationId: AdicionarTrilha
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdicionarTrilhaRequest'
        required: true
      responses:
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdicionarTrilhaResponse'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
...
  '/api/v1/eventos/{id}/trilhas/{trilhaId}':
    delete:
      tags:
        - Eventos
      summary: Exclui logicamente uma trilha
      description: O evento deve manter ao menos uma trilha. **Perfil exigido:** Administrador ou Organizador.
      operationId: ExcluirTrilha
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: trilhaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    put:
      tags:
        - Eventos
      summary: Atualiza uma trilha do evento
      description: Atualiza identificação e disponibilidade da trilha. **Perfil exigido:** Administrador ou Organizador.
      operationId: AtualizarTrilha
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: trilhaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarTrilhaRequest'
        required: true
      responses:
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtualizarTrilhaResponse'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
...
components:
  schemas:
...
    AdicionarTrilhaRequest:
      required:
        - trilhaNome
        - trilhaDescricao
        - trilhaCor
      type: object
      properties:
        trilhaNome:
          type: string
        trilhaDescricao:
          type:
            - 'null'
            - string
        trilhaCor:
          type:
            - 'null'
            - string
    AdicionarTrilhaResponse:
      required:
        - id
        - eventoId
        - trilhaNome
        - trilhaDescricao
        - trilhaCor
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventoId:
          type: string
          format: uuid
        trilhaNome:
          type: string
        trilhaDescricao:
          type:
            - 'null'
            - string
        trilhaCor:
          type:
            - 'null'
            - string
...
    AtualizarTrilhaRequest:
      required:
        - trilhaNome
        - trilhaDescricao
        - trilhaCor
        - estaAtivo
      type: object
      properties:
        trilhaNome:
          type: string
        trilhaDescricao:
          type:
            - 'null'
            - string
        trilhaCor:
          type:
            - 'null'
            - string
        estaAtivo:
          type: boolean
    AtualizarTrilhaResponse:
      required:
        - id
        - eventoId
        - trilhaNome
        - trilhaDescricao
        - trilhaCor
        - estaAtivo
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventoId:
          type: string
          format: uuid
        trilhaNome:
          type: string
        trilhaDescricao:
          type:
            - 'null'
            - string
        trilhaCor:
          type:
            - 'null'
            - string
        estaAtivo:
          type: boolean
...
    HttpValidationProblemDetails:
      type: object
      properties:
        type:
          type:
            - 'null'
            - string
        title:
          type:
            - 'null'
            - string
        status:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        detail:
          type:
            - 'null'
            - string
        instance:
          type:
            - 'null'
            - string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ListarTrilhasItemResponse:
      required:
        - id
        - eventoId
        - trilhaNome
        - trilhaDescricao
        - trilhaCor
        - estaAtivo
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventoId:
          type: string
          format: uuid
        trilhaNome:
          type: string
        trilhaDescricao:
          type:
            - 'null'
            - string
        trilhaCor:
          type:
            - 'null'
            - string
        estaAtivo:
          type: boolean
...
    ProblemDetails:
      type: object
      properties:
        type:
          type:
            - 'null'
            - string
        title:
          type:
            - 'null'
            - string
        status:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        detail:
          type:
            - 'null'
            - string
        instance:
          type:
            - 'null'
            - string
...
  securitySchemes:
    Bearer:
      type: http
      description: Informe apenas o token (o prefixo `Bearer` é adicionado automaticamente).
      scheme: bearer
      bearerFormat: JWT
...